### Enhancement
For some random reason, some versions of VLC don't save .m3u playlists correctly. In fact you then try and open the playlist in VLC it doesn't even load... None-the-less, the file paths are present in the file, so the playlist can still be parsed. The parser is present in this version.

### Bug Analysis - Auto Upload of Playlists

A lot of people are getting 400 - Bad Request when the application tries to create a YTM playlist. I've no idea why this is and can't replicate it, and have been trying to figure it out. This version will provide additional diagnostic data when that error is encountered so I can troubleshoot the issue.